Use fake transport to verify uow startup check
