+++ src/NServiceBus.AcceptanceTests/FakeTransport/FakeTransport.cs
@@ -1,10 +1,5 @@
 ﻿namespace NServiceBus.AcceptanceTests.FakeTransport
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Threading.Tasks;
-    using NServiceBus.Routing;
     using Settings;
     using Transports;
 
@@ -18,45 +13,5 @@
         public override bool RequiresConnectionString => false;
 
         public override string ExampleConnectionStringForErrorMessage => null;
-    }
-
-    public class FakeTransportInfrastructure : TransportInfrastructure
-    {
-        readonly ReadOnlySettings settings;
-
-        public FakeTransportInfrastructure(ReadOnlySettings settings)
-        {
-            this.settings = settings;
-        }
-
-        public override EndpointInstance BindToLocalEndpoint(EndpointInstance instance)
-        {
-            return instance;
-        }
-
-        public override string ToTransportAddress(LogicalAddress logicalAddress)
-        {
-            return logicalAddress.ToString();
-        }
-
-        public override TransportReceiveInfrastructure ConfigureReceiveInfrastructure()
-        {
-            return new TransportReceiveInfrastructure(() => new FakeReceiver(settings.Get<Exception>()), () => new FakeQueueCreator(), () => Task.FromResult(StartupCheckResult.Success));
-        }
-
-        public override TransportSendInfrastructure ConfigureSendInfrastructure()
-        {
-            return new TransportSendInfrastructure(() => new FakeDispatcher(), () => Task.FromResult(StartupCheckResult.Success));
-        }
-
-        public override TransportSubscriptionInfrastructure ConfigureSubscriptionInfrastructure()
-        {
-            throw new NotImplementedException();
-        }
-
-        public override IEnumerable<Type> DeliveryConstraints { get; } = Enumerable.Empty<Type>();
-        public override TransportTransactionMode TransactionMode { get; } = TransportTransactionMode.ReceiveOnly;
-        public override OutboundRoutingPolicy OutboundRoutingPolicy { get; } = new OutboundRoutingPolicy(OutboundRoutingType.Unicast, OutboundRoutingType.Unicast, OutboundRoutingType.Unicast);
-        
     }
 }
