+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -69,6 +69,7 @@
     <Compile Include="FakeTransport\FakeQueueCreator.cs" />
     <Compile Include="FakeTransport\FakeReceiver.cs" />
     <Compile Include="FakeTransport\FakeTransport.cs" />
+    <Compile Include="FakeTransport\FakeTransportInfrastructure.cs" />
     <Compile Include="FakeTransport\FakeTransportSetingsExtensions.cs" />
     <Compile Include="FakeTransport\ProcessingOptimizations\When_using_concurrency_limit.cs" />
     <Compile Include="Hosting\When_overriding_local_addresses.cs" />
